+++ src/components/HotspotsMap/HexHotspots.tsx
@@ -1,13 +1,11 @@
-import animalHash from "angry-purple-tiger"
 import clsx from "clsx"
-import { HeliumIotIcon } from "../icons/HeliumIotIcon"
-import { HeliumMobileIcon } from "../icons/HeliumMobileIcon"
+import { HexHotSpotItem } from "./HexHotspotItem"
 
 interface SmallCell {
   cell_id: string
 }
 
-interface Hotspot {
+export interface Hotspot {
   hotspot_id: string
   active: boolean
   cells: SmallCell[]
@@ -77,43 +75,9 @@ export async function HexHotspots({ hexId }: { hexId: string }) {
               role="list"
               className="z-0 flex-1 divide-y divide-gray-200 overflow-y-auto dark:divide-white/10"
             >
-              {groupedList[group].map((hotspot) => {
-                const hotspotName = animalHash(hotspot.hotspot_id)
-                const hasSmallCells = hotspot.cells.length > 0
-                const Avatar = hasSmallCells ? HeliumMobileIcon : HeliumIotIcon
-                const subtitle = hasSmallCells
-                  ? `${hotspot.cells.length} small cell${
-                      hotspot.cells.length === 1 ? "" : "s"
-                    }`
-                  : "IoT Hotspot"
-                return (
-                  <li key={hotspot.hotspot_id}>
-                    <div className="group relative flex items-center px-2 py-3">
-                      <a
-                        href={`https://app.hotspotty.net/hotspots/${hotspot.hotspot_id}/rewards`}
-                        target="_blank"
-                        className="-m-1 block flex-1 p-1"
-                      >
-                        <div
-                          className="absolute inset-0 group-hover:bg-zinc-300/30 dark:group-hover:bg-zinc-700/30"
-                          aria-hidden="true"
-                        />
-                        <div className="relative flex min-w-0 flex-1 items-center gap-4">
-                          <Avatar className="inline-block h-8 w-8 flex-shrink-0" />
-                          <div className="truncate">
-                            <p className="truncate text-sm font-medium leading-5 text-gray-900 dark:text-zinc-100">
-                              {hotspotName}
-                            </p>
-                            <p className="truncate text-xs text-gray-600 dark:text-zinc-300">
-                              {subtitle}
-                            </p>
-                          </div>
-                        </div>
-                      </a>
-                    </div>
-                  </li>
-                )
-              })}
+              {groupedList[group].map((hotspot) => (
+                <HexHotSpotItem key={hotspot.hotspot_id} hotspot={hotspot} />
+              ))}
             </ul>
           </div>
         )
